+++ examples/ts/src/test.ts
@@ -1,0 +1,5 @@
+// This is copied to the run folder for testing
+
+let json : Json = require("std:json");
+
+src.entity.inventory.set(0, {id:"minecraft:dirt", count:1, nbt:{test:"Hello"}});