+++ luadatapacktypes/src/json.d.ts
@@ -1,0 +1,6 @@
+/** @noSelfInFile **/
+
+interface Json {
+    encode(this : void, obj : any) : string;
+    decode(this : void, obj : string) : any;
+}